Move DoodleVideo related components into one folder, rename them, and deprecate unused components
