+++ src/lib/DoodleVideo/LegacyHDSingleBoard.svelte
@@ -118,9 +118,9 @@
 </div>
 
 <script>
-  import { connectTwoPoints, drawStroke, renderBackground } from '../helpers/canvas.js'
+  import { connectTwoPoints, drawStroke, renderBackground } from '../../helpers/canvas.js'
   import { onMount, onDestroy, createEventDispatcher } from 'svelte'
-  import { maxAvailableWidth, maxAvailableHeight, assumedCanvasWidth, user } from '../store.js' // note `canvasWidth` was misleading
+  import { maxAvailableWidth, maxAvailableHeight, assumedCanvasWidth, user } from '../../store.js' // note `canvasWidth` was misleading
   import Button, { Label } from '@smui/button'
   import BaseTransparentButton from '$lib/BaseTransparentButton.svelte'
